+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/env/start.py
@@ -401,7 +401,7 @@ def _get_agent_ver(ctx, python, agent):
     if os.path.isdir(legacy_fallback):
         legacy_fallback = ''
 
-    fallback = os.getenv('DDEV_E2E_AGENT', legacy_fallback)
+    fallback = os.getenv('DDEV_E2E_AGENT') or legacy_fallback
     # DDEV_E2E_AGENT_PY2 overrides DDEV_E2E_AGENT when starting a Python 2 environment
     if python == 2 and os.getenv('DDEV_E2E_AGENT_PY2'):
         fallback = os.getenv('DDEV_E2E_AGENT_PY2')