+++ mysql/datadog_checks/mysql/statement_samples.py
@@ -199,6 +199,7 @@ class MySQLStatementSamples(DBMAsyncJob):
         self._connection_args = connection_args
         self._last_check_run = 0
         self._db = None
+        self._check = check
         self._configured_collection_interval = self._config.statement_samples_config.get('collection_interval', -1)
         self._events_statements_row_limit = self._config.statement_samples_config.get(
             'events_statements_row_limit', 5000
@@ -479,7 +480,7 @@ class MySQLStatementSamples(DBMAsyncJob):
         """
         with closing(self._get_db_connection().cursor()) as cursor:
             self._cursor_run(cursor, ENABLED_STATEMENTS_CONSUMERS_QUERY)
-            return set([r[0] for r in cursor.fetchall()])
+            return {r[0] for r in cursor.fetchall()}
 
     def _enable_events_statements_consumers(self):
         """