+++ postgres/tests/common.py
@@ -10,6 +10,7 @@ from datadog_checks.base.stubs.aggregator import normalize_tags
 from datadog_checks.dev import get_docker_hostname
 from datadog_checks.dev.docker import get_container_ip
 from datadog_checks.postgres.util import (
+    NEWER_14_METRICS,
     QUERY_PG_REPLICATION_SLOTS,
     QUERY_PG_STAT_WAL_RECEIVER,
     REPLICATION_STATS_METRICS,
@@ -20,8 +21,11 @@ from datadog_checks.postgres.version_utils import VersionUtils
 HOST = get_docker_hostname()
 PORT = '5432'
 PORT_REPLICA = '5433'
+PORT_REPLICA2 = '5434'
 USER = 'datadog'
+USER_ADMIN = 'dd_admin'
 PASSWORD = 'datadog'
+PASSWORD_ADMIN = 'dd_admin'
 DB_NAME = 'datadog_test'
 POSTGRES_VERSION = os.environ.get('POSTGRES_VERSION', None)
 POSTGRES_IMAGE = "alpine"
@@ -59,6 +63,14 @@ COMMON_METRICS = [
 
 DBM_MIGRATED_METRICS = [
     'postgresql.connections',
+]
+
+CONFLICT_METRICS = [
+    'postgresql.conflicts.tablespace',
+    'postgresql.conflicts.lock',
+    'postgresql.conflicts.snapshot',
+    'postgresql.conflicts.bufferpin',
+    'postgresql.conflicts.deadlock',
 ]
 
 COMMON_BGW_METRICS = [
@@ -108,13 +120,16 @@ def check_common_metrics(aggregator, expected_tags, count=1):
         db_tags = expected_tags + ['db:{}'.format(db)]
         for name in COMMON_METRICS:
             aggregator.assert_metric(name, count=count, tags=db_tags)
+        if POSTGRES_VERSION is None or float(POSTGRES_VERSION) >= 14.0:
+            for (metric_name, _) in NEWER_14_METRICS.values():
+                aggregator.assert_metric(metric_name, count=count, tags=db_tags)
 
 
 def check_db_count(aggregator, expected_tags, count=1):
     aggregator.assert_metric(
         'postgresql.table.count', value=5, count=count, tags=expected_tags + ['db:{}'.format(DB_NAME), 'schema:public']
     )
-    aggregator.assert_metric('postgresql.db.count', value=5, count=1)
+    aggregator.assert_metric('postgresql.db.count', value=106, count=1)
 
 
 def check_connection_metrics(aggregator, expected_tags, count=1):
@@ -191,6 +206,15 @@ def check_replication_delay(aggregator, metrics_cache, expected_tags, count=1):
     replication_metrics = metrics_cache.get_replication_metrics(VersionUtils.parse_version(POSTGRES_VERSION), False)
     for (metric_name, _) in replication_metrics.values():
         aggregator.assert_metric(metric_name, count=count, tags=expected_tags)
+
+
+def check_conflict_metrics(aggregator, expected_tags, count=1):
+    if float(POSTGRES_VERSION) < 9.1:
+        return
+    for db in COMMON_DBS:
+        db_tags = expected_tags + ['db:{}'.format(db)]
+        for name in CONFLICT_METRICS:
+            aggregator.assert_metric(name, count=count, tags=db_tags)
 
 
 def check_bgw_metrics(aggregator, expected_tags, count=1):