+++ postgres/tests/test_pg_replication.py
@@ -7,11 +7,10 @@ import psycopg2
 import pytest
 
 from .common import (
-    DB_NAME,
-    HOST,
     assert_metric_at_least,
     check_bgw_metrics,
     check_common_metrics,
+    check_conflict_metrics,
     check_connection_metrics,
     check_db_count,
     check_replication_delay,
@@ -21,6 +20,29 @@ from .common import (
 from .utils import requires_over_10
 
 pytestmark = [pytest.mark.integration, pytest.mark.usefixtures('dd_environment')]
+
+
+def _get_superconn(db_instance, application_name='test'):
+    conn = psycopg2.connect(
+        host=db_instance['host'],
+        dbname=db_instance['dbname'],
+        user='postgres',
+        password='datad0g',
+        port=db_instance['port'],
+        application_name=application_name,
+    )
+    return conn
+
+
+def _wait_for_value(db_instance, lower_threshold, query):
+    value = 0
+    with _get_superconn(db_instance) as conn:
+        conn.autocommit = True
+        while value <= lower_threshold:
+            with conn.cursor() as cur:
+                cur.execute(query)
+                value = cur.fetchall()[0][0]
+            time.sleep(0.1)
 
 
 @requires_over_10
@@ -36,15 +58,16 @@ def test_common_replica_metrics(aggregator, integration_check, metrics_cache_rep
     check_slru_metrics(aggregator, expected_tags=expected_tags)
     check_replication_delay(aggregator, metrics_cache_replica, expected_tags=expected_tags)
     check_wal_receiver_metrics(aggregator, expected_tags=expected_tags + ['status:streaming'])
+    check_conflict_metrics(aggregator, expected_tags=expected_tags)
 
     aggregator.assert_all_metrics_covered()
 
 
 @requires_over_10
-def test_wal_receiver_metrics(aggregator, integration_check, pg_replica_instance):
+def test_wal_receiver_metrics(aggregator, integration_check, pg_instance, pg_replica_instance):
     check = integration_check(pg_replica_instance)
     expected_tags = pg_replica_instance['tags'] + ['port:{}'.format(pg_replica_instance['port']), 'status:streaming']
-    with psycopg2.connect(host=HOST, dbname=DB_NAME, user="postgres", password="datad0g") as conn:
+    with _get_superconn(pg_instance) as conn:
         with conn.cursor() as cur:
             # Ask for a new txid to force a WAL change
             cur.execute('select txid_current();')
@@ -78,3 +101,88 @@ def test_wal_receiver_metrics(aggregator, integration_check, pg_replica_instance
         tags=expected_tags,
         count=1,
     )
+
+
+@requires_over_10
+def test_conflicts_lock(aggregator, integration_check, pg_instance, pg_replica_instance2):
+    check = integration_check(pg_replica_instance2)
+    expected_tags = pg_replica_instance2['tags'] + ['port:{}'.format(pg_replica_instance2['port']), 'db:datadog_test']
+
+    replica_con = _get_superconn(pg_replica_instance2)
+    replica_cur = replica_con.cursor()
+    replica_cur.execute('BEGIN;')
+    replica_cur.execute('select * from persons;')
+    replica_cur.fetchall()
+
+    with _get_superconn(pg_instance) as conn:
+        conn.autocommit = True
+        with conn.cursor() as cur:
+            cur.execute('update persons SET personid = 1 where personid = 1;')
+            cur.execute('vacuum full persons')
+    _wait_for_value(
+        pg_replica_instance2,
+        lower_threshold=0,
+        query="select confl_lock from pg_stat_database_conflicts where datname='datadog_test';",
+    )
+
+    check.check(pg_replica_instance2)
+    aggregator.assert_metric('postgresql.conflicts.lock', value=1, tags=expected_tags)
+
+
+@requires_over_10
+def test_conflicts_snapshot(aggregator, integration_check, pg_instance, pg_replica_instance2):
+    check = integration_check(pg_replica_instance2)
+    expected_tags = pg_replica_instance2['tags'] + ['port:{}'.format(pg_replica_instance2['port']), 'db:datadog_test']
+
+    replica2_con = _get_superconn(pg_replica_instance2)
+    replica2_cur = replica2_con.cursor()
+    replica2_cur.execute('BEGIN TRANSACTION ISOLATION LEVEL REPEATABLE READ;')
+    replica2_cur.execute('select * from persons;')
+
+    with _get_superconn(pg_instance) as conn:
+        conn.autocommit = True
+        with conn.cursor() as cur:
+            cur.execute('update persons SET personid = 1 where personid = 1;')
+            time.sleep(0.2)
+            cur.execute('vacuum verbose persons;')
+
+    _wait_for_value(
+        pg_replica_instance2,
+        lower_threshold=0,
+        query="select confl_snapshot from pg_stat_database_conflicts where datname='datadog_test';",
+    )
+    check.check(pg_replica_instance2)
+    aggregator.assert_metric('postgresql.conflicts.snapshot', value=1, tags=expected_tags)
+
+
+@requires_over_10
+def test_conflicts_bufferpin(aggregator, integration_check, pg_instance, pg_replica_instance2):
+    check = integration_check(pg_replica_instance2)
+    expected_tags = pg_replica_instance2['tags'] + ['port:{}'.format(pg_replica_instance2['port']), 'db:datadog_test']
+
+    with _get_superconn(pg_instance) as conn:
+        with conn.cursor() as cur:
+            cur.execute('BEGIN;')
+            cur.execute("INSERT INTO persons VALUES (3,'t','t','t');")
+            cur.execute('ROLLBACK;')
+
+    replica2_con = _get_superconn(pg_replica_instance2)
+    replica2_cur = replica2_con.cursor()
+    replica2_cur.execute('BEGIN;')
+    replica2_cur.execute('DECLARE cursor1 CURSOR FOR SELECT * FROM persons')
+    replica2_cur.execute('FETCH FORWARD FROM cursor1')
+    replica2_cur.fetchall()
+
+    with _get_superconn(pg_instance) as conn:
+        conn.autocommit = True
+        with conn.cursor() as cur:
+            cur.execute('vacuum verbose persons;')
+
+    _wait_for_value(
+        pg_replica_instance2,
+        lower_threshold=0,
+        query="select confl_bufferpin from pg_stat_database_conflicts where datname='datadog_test';",
+    )
+
+    check.check(pg_replica_instance2)
+    aggregator.assert_metric('postgresql.conflicts.bufferpin', value=1, tags=expected_tags)