+++ snmp/tests/test_e2e_core.py
@@ -2,9 +2,9 @@
 # All rights reserved
 # Licensed under Simplified BSD License (see LICENSE)
 import pytest
-from tests.common import SNMP_CONTAINER_NAME
 
 from datadog_checks.dev.docker import get_container_ip
+from tests.common import SNMP_CONTAINER_NAME
 
 from . import common, metrics
 
@@ -312,7 +312,7 @@ def test_e2e_meraki_cloud_controller(dd_agent_check):
     aggregator.assert_metric(
         'snmp.interface.status',
         metric_type=aggregator.GAUGE,
-        tags=if_tags + ['interface_index:11', 'status:warning'],
+        tags=if_tags + ['interface_index:11', 'status:warning', 'admin_status:down', 'oper_status:lower_layer_down'],
         value=1,
     )
     aggregator.assert_all_metrics_covered()